Evaluation/requirements.txt
tqdm==4.66.5
openai==1.51.2
python-dotenv==1.0.1